refactored organization page structure into pageand component
